requirements.txt
# Core data science packages
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning packages
scikit-learn>=1.3.0
xgboost>=1.7.0

# Deep learning packages
torch>=2.0.0
transformers>=4.30.0

# Bioinformatics packages
biopython>=1.81
scipy>=1.10.0

# Sequence analysis and utilities
logomaker>=0.8
pyfaidx>=0.7.0

# Development tools
jupyter>=1.0.0
ipython>=8.0.0
tqdm>=4.65.0

# Data visualization
plotly>=5.15.0